+++ src/avl.h
@@ -9,7 +9,7 @@ typedef struct AvlNode AvlNode;
 
 struct AvlNode
 {
-    int32    key;
+    int64    key;
     int64    value;
     AvlNode  *left;
     AvlNode  *right;
@@ -17,7 +17,7 @@ struct AvlNode
 };
 
 AvlNode* istore_make_empty(AvlNode *t);
-AvlNode* is_tree_find(int32 key, AvlNode *t);
-AvlNode* is_tree_insert(AvlNode *t, int32 key, int64 value);
+AvlNode* is_tree_find(int64 key, AvlNode *t);
+AvlNode* is_tree_insert(AvlNode *t, int64 key, int64 value);
 int is_tree_length(AvlNode* t);
 #endif // ISTORE_AVL_H
